+++ packages/widgets-manager/src/configureStore.ts
@@ -1,18 +1,11 @@
-import { createEpicMiddleware } from 'redux-observable';
 import { createLogger } from 'redux-logger';
 import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension/logOnly';
 import { rootReducer } from './reducers';
-import { rootEpic } from './epics';
 
-export const configureStore = (initialState: any, dependencies: any) => {
-  const epicMiddleware = createEpicMiddleware({ dependencies });
-
+export const configureStore = (initialState: any) => {
   // Create the store with three middlewares (the order here is important)
-  // 1. epicMiddleware: Makes redux-observables work
-  const middlewares = [
-    epicMiddleware,
-  ];
+  const middlewares = [];
 
   const enhancers: any[] = [];
 
@@ -48,9 +41,6 @@ export const configureStore = (initialState: any, dependencies: any) => {
     initialState,
     composeEnhancers(...enhancers)
   );
-
-  /* Finally attach epics to store */
-  epicMiddleware.run(rootEpic);
 
   return store;
 }