+++ packages/widgets-manager/CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Hotfix [2.2.1] - 2020-01-05
+## Hotfix [2.2.1, 2.2.2] - 2020-01-05
 ### Fixed
 - Disable the Sync of firebase with redux -> Make a TODO to figure out corner
 cases and enable in the future
+- Disable Epics to disable any firebase side effect like mounting apps
 
 ## Hotfix [2.2.0] - 2019-12-12
 ### Fixed